+++ unicorn_binance_websocket_api/restclient.py
@@ -39,6 +39,7 @@
 # IN THE SOFTWARE.
 
 from unicorn_binance_rest_api import BinanceRestApiManager
+from unicorn_binance_rest_api.exceptions import UnknownExchange
 from typing import Optional, Union, Tuple
 import logging
 import requests
@@ -87,16 +88,26 @@ class BinanceWebSocketApiRestclient(object):
         """
         Init UBRA if necessary.
 
+        :return: `True` if `self.ubra` is ready to use, `False` if the installed UBRA version
+                 does not (yet) know `self.exchange` - callers must check this and bail out
+                 gracefully instead of letting `UnknownExchange` propagate uncaught.
         """
         if self.ubra is None:
             logger.debug(f"Init UBRA for UBWA restclient.")
-            self.ubra = BinanceRestApiManager(debug=self.debug,
-                                              disable_colorama=self.disable_colorama,
-                                              exchange=self.exchange,
-                                              socks5_proxy_server=self.socks5_proxy_server,
-                                              socks5_proxy_user=self.socks5_proxy_user,
-                                              socks5_proxy_pass=self.socks5_proxy_pass,
-                                              warn_on_update=self.warn_on_update)
+            try:
+                self.ubra = BinanceRestApiManager(debug=self.debug,
+                                                  disable_colorama=self.disable_colorama,
+                                                  exchange=self.exchange,
+                                                  socks5_proxy_server=self.socks5_proxy_server,
+                                                  socks5_proxy_user=self.socks5_proxy_user,
+                                                  socks5_proxy_pass=self.socks5_proxy_pass,
+                                                  warn_on_update=self.warn_on_update)
+            except UnknownExchange as error_msg:
+                logger.critical(f"BinanceWebSocketApiRestclient._init_ubra() - The installed "
+                                f"`unicorn-binance-rest-api` version does not support "
+                                f"exchange='{self.exchange}' yet - please upgrade UBRA - "
+                                f"error_msg: {error_msg}")
+                return False
         return True
 
     def delete_listen_key(self, stream_id=None) -> Tuple[Union[str, None], Union[dict, None]]:
@@ -116,7 +127,8 @@ class BinanceWebSocketApiRestclient(object):
 
         try:
             with (self.threading_lock):
-                self._init_ubra()
+                if self._init_ubra() is False:
+                    return None, None
 
                 try:
                     kwargs = {'api_key': self.stream_list[stream_id]['api_key'],
@@ -152,6 +164,12 @@ class BinanceWebSocketApiRestclient(object):
                     result = self.ubra.futures_coin_stream_close(listenKey=self.stream_list[stream_id]['listen_key'],
                                                                  throw_exception=False,
                                                                  **kwargs)
+                elif self.exchange == "binance.com-portfolio_margin":
+                    if self.restful_base_uri is not None:
+                        self.ubra.PAPI_URL = self.restful_base_uri
+                    result = self.ubra.portfolio_margin_stream_close(listenKey=self.stream_list[stream_id]['listen_key'],
+                                                                     throw_exception=False,
+                                                                     **kwargs)
                 else:
                     if self.restful_base_uri is not None:
                         self.ubra.API_URL = self.restful_base_uri
@@ -196,7 +214,8 @@ class BinanceWebSocketApiRestclient(object):
             return None, None
 
         with (self.threading_lock):
-            self._init_ubra()
+            if self._init_ubra() is False:
+                return None, None
 
             try:
                 kwargs = {'api_key': self.stream_list[stream_id]['api_key'],
@@ -257,6 +276,17 @@ class BinanceWebSocketApiRestclient(object):
                     logger.critical(f"BinanceWebSocketApiRestclient.get_listen_key() - error: 8 - "
                                     f"error_msg: {error_msg} - Can not acquire listen_key for coin futures!!")
                     return None, None
+            elif self.exchange == "binance.com-portfolio_margin":
+                try:
+                    if self.restful_base_uri is not None:
+                        self.ubra.PAPI_URL = self.restful_base_uri
+                    response = self.ubra.portfolio_margin_stream_get_listen_key(output="raw_data",
+                                                                                throw_exception=False,
+                                                                                **kwargs)
+                except AttributeError as error_msg:
+                    logger.critical(f"BinanceWebSocketApiRestclient.get_listen_key() - error: 8 - "
+                                    f"error_msg: {error_msg} - Can not acquire listen_key for portfolio margin!!")
+                    return None, None
             else:
                 try:
                     if self.restful_base_uri is not None:
@@ -296,7 +326,8 @@ class BinanceWebSocketApiRestclient(object):
             return None, None
 
         with (self.threading_lock):
-            self._init_ubra()
+            if self._init_ubra() is False:
+                return None, None
 
             try:
                 kwargs = {'api_key': self.stream_list[stream_id]['api_key'],
@@ -332,6 +363,12 @@ class BinanceWebSocketApiRestclient(object):
                 result = self.ubra.futures_stream_keepalive(listenKey=self.stream_list[stream_id]['listen_key'],
                                                             throw_exception=False,
                                                             **kwargs)
+            elif self.exchange == "binance.com-portfolio_margin":
+                if self.restful_base_uri is not None:
+                    self.ubra.PAPI_URL = self.restful_base_uri
+                result = self.ubra.portfolio_margin_stream_keepalive(listenKey=self.stream_list[stream_id]['listen_key'],
+                                                                     throw_exception=False,
+                                                                     **kwargs)
             else:
                 if self.restful_base_uri is not None:
                     self.ubra.API_URL = self.restful_base_uri